ai-weight-estimation/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Computer Vision
albumentations>=1.3.0
imgaug>=0.4.0
scikit-image>=0.21.0

# Tracking & Re-ID
filterpy>=1.4.5
scipy>=1.11.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
tensorflow>=2.13.0  # Pour TensorFlow Lite

# API & Integration
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0  # Pour synchronisation avec backend

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
